+++ src/tunable.h
@@ -3,8 +3,8 @@
 #include "types.h"
 
 constexpr double CPUCT = 2.4;
-constexpr double FPU = 0.5;
-constexpr double TEMPERATURE = 1.0;
-constexpr double EVAL_DIVISOR = 400;
+constexpr float FPU = 0.5;
+constexpr float TEMPERATURE = 1.0;
+constexpr float EVAL_DIVISOR = 400;
 
 constexpr array<int, 7> PIECE_VALUES = {100, 316, 328, 493, 982, 0, 0};
